refactor: Changed directory naming for consistency
